requirements.txt
django
djangorestframework
django-filter
djangorestframework-api-key

